requirements.txt
streamlit
firebase-admin
pandas
plotly
pyrebase4